requirements.txt
python-telegram-bot>=20.0
Pillow>=9.0.0
python-dotenv>=0.19.0
numpy>=1.21.0
opencv-python-headless>=4.8.0  # Using headless version for serverless
aiohttp>=3.8.1
fastapi>=0.68.0
uvicorn>=0.15.0 